refactor: nacl.js
