+++ lib/secretbox.js
@@ -1,0 +1,37 @@
+import {
+    crypto_secretbox_BOXZEROBYTES,
+    crypto_secretbox_KEYBYTES,
+    crypto_secretbox_NONCEBYTES,
+    crypto_secretbox_ZEROBYTES
+} from './constants.js';
+import {checkArrayTypes} from './utils.js';
+import {crypto_secretbox, crypto_secretbox_open} from './crypto.js';
+
+function checkLengths(k, n) {
+    if (k.length !== crypto_secretbox_KEYBYTES) throw new Error('bad key size');
+    if (n.length !== crypto_secretbox_NONCEBYTES) throw new Error('bad nonce size');
+}
+export function secretbox(msg, nonce, key) {
+    checkArrayTypes(msg, nonce, key);
+    checkLengths(key, nonce);
+    var m = new Uint8Array(crypto_secretbox_ZEROBYTES + msg.length);
+    var c = new Uint8Array(m.length);
+    for (var i = 0; i < msg.length; i++) m[i+crypto_secretbox_ZEROBYTES] = msg[i];
+    crypto_secretbox(c, m, m.length, nonce, key);
+    return c.subarray(crypto_secretbox_BOXZEROBYTES);
+}
+
+secretbox.open = function(box, nonce, key) {
+    checkArrayTypes(box, nonce, key);
+    checkLengths(key, nonce);
+    var c = new Uint8Array(crypto_secretbox_BOXZEROBYTES + box.length);
+    var m = new Uint8Array(c.length);
+    for (var i = 0; i < box.length; i++) c[i+crypto_secretbox_BOXZEROBYTES] = box[i];
+    if (c.length < 32) return null;
+    if (crypto_secretbox_open(m, c, c.length, nonce, key) !== 0) return null;
+    return m.subarray(crypto_secretbox_ZEROBYTES);
+};
+
+secretbox.keyLength = crypto_secretbox_KEYBYTES;
+secretbox.nonceLength = crypto_secretbox_NONCEBYTES;
+secretbox.overheadLength = crypto_secretbox_BOXZEROBYTES;
