+++ test/02-randombytes.quick.js
@@ -1,12 +1,12 @@
-var nacl = (typeof window !== 'undefined') ? window.nacl : require('../' + (process.env.NACL_SRC || 'nacl.min.js'));
-nacl.util = require('tweetnacl-util');
-var test = require('tape');
+import * as nacl from '../nacl.js'
+import test from 'tape'
+import util from 'tweetnacl-util'
 
 test('nacl.randomBytes', function(t) {
   t.plan(1);
   var set = {}, s, i;
   for (i = 0; i < 10000; i++) {
-    s = nacl.util.encodeBase64(nacl.randomBytes(32));
+    s = util.encodeBase64(nacl.randomBytes(32));
     if (set[s]) {
       t.fail('duplicate random sequence! ', s);
       return;
