+++ test/06-box.quick.js
@@ -1,8 +1,8 @@
-var nacl = (typeof window !== 'undefined') ? window.nacl : require('../' + (process.env.NACL_SRC || 'nacl.min.js'));
-nacl.util = require('tweetnacl-util');
-var test = require('tape');
+import * as nacl from '../nacl.js'
+import test from 'tape'
+import util from 'tweetnacl-util'
 
-var enc = nacl.util.encodeBase64;
+var enc = util.encodeBase64;
 
 test('nacl.box.keyPair', function(t) {
   var keys = nacl.box.keyPair();
@@ -25,14 +25,14 @@ test('nacl.box and nacl.box.open', function(t) {
   var serverKeys = nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
   for (var i = 0; i < nonce.length; i++) nonce[i] = (32+i) & 0xff;
-  var msg = nacl.util.decodeUTF8('message to encrypt');
+  var msg = util.decodeUTF8('message to encrypt');
   var clientBox = nacl.box(msg, nonce, serverKeys.publicKey, clientKeys.secretKey);
   var clientMsg = nacl.box.open(clientBox, nonce, clientKeys.publicKey, serverKeys.secretKey);
-  t.equal(nacl.util.encodeUTF8(clientMsg), nacl.util.encodeUTF8(msg));
+  t.equal(util.encodeUTF8(clientMsg), util.encodeUTF8(msg));
   var serverBox = nacl.box(msg, nonce, clientKeys.publicKey, serverKeys.secretKey);
   t.equal(enc(clientBox), enc(serverBox));
   var serverMsg = nacl.box.open(serverBox, nonce, serverKeys.publicKey, clientKeys.secretKey);
-  t.equal(nacl.util.encodeUTF8(serverMsg), nacl.util.encodeUTF8(msg));
+  t.equal(util.encodeUTF8(serverMsg), util.encodeUTF8(msg));
   t.end();
 });
 
@@ -51,10 +51,10 @@ test('nacl.box.open with invalid nonce', function(t) {
   var serverKeys = nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
   for (var i = 0; i < nonce.length; i++) nonce[i] = i & 0xff;
-  var msg = nacl.util.decodeUTF8('message to encrypt');
+  var msg = util.decodeUTF8('message to encrypt');
   var box = nacl.box(msg, nonce, clientKeys.publicKey, serverKeys.secretKey);
-  t.equal(nacl.util.encodeUTF8(nacl.box.open(box, nonce, serverKeys.publicKey, clientKeys.secretKey)),
-          nacl.util.encodeUTF8(msg));
+  t.equal(util.encodeUTF8(nacl.box.open(box, nonce, serverKeys.publicKey, clientKeys.secretKey)),
+          util.encodeUTF8(msg));
   nonce[0] = 255;
   t.equal(nacl.box.open(box, nonce, serverKeys.publicKey, clientKeys.secretKey), null);
   t.end();
@@ -64,12 +64,12 @@ test('nacl.box.open with invalid keys', function(t) {
   var clientKeys = nacl.box.keyPair();
   var serverKeys = nacl.box.keyPair();
   var nonce = new Uint8Array(nacl.box.nonceLength);
-  var msg = nacl.util.decodeUTF8('message to encrypt');
+  var msg = util.decodeUTF8('message to encrypt');
   var box = nacl.box(msg, nonce, clientKeys.publicKey, serverKeys.secretKey);
-  t.equal(nacl.util.encodeUTF8(nacl.box.open(box, nonce, serverKeys.publicKey, clientKeys.secretKey)),
-          nacl.util.encodeUTF8(msg));
-  t.equal(nacl.util.encodeUTF8(nacl.box.open(box, nonce, clientKeys.publicKey, serverKeys.secretKey)),
-          nacl.util.encodeUTF8(msg));
+  t.equal(util.encodeUTF8(nacl.box.open(box, nonce, serverKeys.publicKey, clientKeys.secretKey)),
+          util.encodeUTF8(msg));
+  t.equal(util.encodeUTF8(nacl.box.open(box, nonce, clientKeys.publicKey, serverKeys.secretKey)),
+          util.encodeUTF8(msg));
   var badPublicKey = new Uint8Array(nacl.box.publicKeyLength);
   t.equal(nacl.box.open(box, nonce, badPublicKey, clientKeys.secretKey), null);
   var badSecretKey = new Uint8Array(nacl.box.secretKeyLength);
