+++ test/data/box.random.js
@@ -1,4 +1,4 @@
-module.exports = [
+export default [
 [
 "vSwMiFeDLVceYF4LXzDB45w7O6jMzTF28yiYpe27Sk8=",
 "fRU+1AfJ8wFqZcRKlwzJAMmFD42dpLFOHUrCFRUefKE=",
