+++ test/data/hash.random.js
@@ -1,4 +1,4 @@
-module.exports = [
+export default [
 [
 "",
 "z4PhNX7vuL3xVChQ1m2AB9Yg5AULVxXcg/SpIdNs6c5H0NE8XYXysP+DGNKHfuwvY7kxvUdBeoGlODJ6+SfaPg=="
