+++ test/data/onetimeauth.spec.js
@@ -1,4 +1,4 @@
-module.exports = [
+export default [
   {
     m: new Uint8Array([72,101,108,108,111,32,119,111,114,108,100,33]),
     k: new Uint8Array([116, 104, 105, 115, 32, 105, 115, 32, 51, 50, 45, 98, 121, 116, 101,
